requirements.txt
sqlalchemy
pydantic
passlib[bcrypt]
PyJWT
shadai